socket_HW/requirements.txt
asyncio==3.4.3
numpy==1.25.2
python-dotenv==1.0.0
websockets==12.0
opencv-contrib-python==4.8.1.78
Pillow==10.1.0
redis==5.0.1
